refactor: SonarLint 검사 및 수정

+++ src/main/java/study/object/ch14/billing/step01/DayOfWeekDiscountRule.java
@@ -25,7 +25,7 @@ public class DayOfWeekDiscountRule {
 
     public Money calculate(DateTimeInterval interval) {
         if (dayOfWeeks.contains(interval.getFrom().getDayOfWeek())) {
-            return amount.times(interval.duration().getSeconds() / duration.getSeconds());
+            return amount.times((double) interval.duration().getSeconds() / duration.getSeconds());
         }
 
         return Money.ZERO;
